@@ -399,7 +399,7 @@ func TestTester(t *testing.T) {
 			name:   "regex grouped variables test",
 			main:   "../../examples/testing/regex/regex.vcl",
 			filter: "*regex.test.vcl",
-			passes: 4,
+			passes: 72,
 		},
 		{
 			name:   "header subfield dealing test",
