@@ -108,7 +108,7 @@ func TestSetRequestHeaderValueOverwrite(t *testing.T) {
 	if diff := cmp.Diff(req.Header, ghttp.Header{
 		"Foo": []string{"abc=123,bar=snafu"},
 	}); diff != "" {
-		t.Errorf(diff)
+		t.Error(diff)
 	}
 }
 
@@ -182,7 +182,7 @@ func TestSetResponseHeaderValueOverwrite(t *testing.T) {
 	if diff := cmp.Diff(resp.Header, ghttp.Header{
 		"Foo": []string{"abc=123,bar=snafu"},
 	}); diff != "" {
-		t.Errorf(diff)
+		t.Error(diff)
 	}
 }
 
