@@ -282,11 +282,11 @@ func (l *Linter) lintInfixExpression(exp *ast.InfixExpression, ctx *context.Cont
 		}
 		return types.BoolType
 	case "~", "!~":
-		// Regex operator could compare only STRING,  IP or ACL type
+		// Regex operator could compare only STRING, IP or ACL type
 		if !expectType(left, types.StringType, types.IPType) {
 			l.Error(InvalidTypeExpression(exp.GetMeta(), left, types.StringType, types.IPType, types.AclType).Match(OPERATOR_CONDITIONAL))
-		} else if !expectType(right, types.StringType, types.AclType) {
-			l.Error(InvalidTypeExpression(exp.GetMeta(), right, types.StringType).Match(OPERATOR_CONDITIONAL))
+		} else if !expectType(right, types.StringType, types.AclType, types.RegexType) {
+			l.Error(InvalidTypeExpression(exp.GetMeta(), right, types.StringType, types.RegexType).Match(OPERATOR_CONDITIONAL))
 		}
 		if expectType(right, types.StringType) && !isLiteralExpression(exp.Right) {
 			l.Error(&LintError{
